Debug

+++ easypap-se/kernel/c/life.c
@@ -49,7 +49,7 @@ void store_change() {
   int tmp_x = 0;
   int tmp_y = 0;
   for (int i = 0; i < DIM/TILE_W; i++) {
-    tmp_x = after_change_x[i];
+    tmp_x = (int)after_change_x[i];
     before_change_x[i] = tmp_x;
   }
   for (int j = 0; j < DIM/TILE_H; j++) {
